extracted applyDefaultConfig and applySectionConfig to seperate file
