+++ app/contextCreator/contextCreator.test.js
@@ -323,95 +323,6 @@ describe('createSolutionPageContext', () => {
   });
 });
 
-describe('applySectionConfig', () => {
-  it('should only return sections defined in the config', () => {
-    const expectedContext = [
-      {
-        id: 'first-section',
-        name: 'First Section',
-        value: 'First Section Value',
-        showTitle: true,
-      },
-    ];
-
-
-    const config = {
-      'first-section': {},
-    };
-
-    const initialContext = [
-      {
-        id: 'first-section',
-        name: 'First Section',
-        value: 'First Section Value',
-      },
-      {
-        id: 'unknown-section',
-        name: 'Unknown Section',
-        value: 'Unknown Section Value',
-      },
-    ];
-
-    const context = applySectionConfig(initialContext, config);
-
-    expect(context).toEqual(expectedContext);
-  });
-
-  it('should decorate a section with the config provided for that section', () => {
-    const expectedContext = [
-      {
-        id: 'first-section',
-        name: 'First Section',
-        value: 'First Section Value',
-        showTitle: false,
-      },
-    ];
-
-    const config = {
-      'first-section': {
-        showTitle: false,
-      },
-    };
-
-    const initialContext = [
-      {
-        id: 'first-section',
-        name: 'First Section',
-        value: 'First Section Value',
-      },
-    ];
-
-    const context = applySectionConfig(initialContext, config);
-
-    expect(context).toEqual(expectedContext);
-  });
-});
-
-describe('applyDefaultConfig', () => {
-  it('show decorate the section with the default config when a config is not provided', () => {
-    const expectedContext = [
-      {
-        id: 'first-section',
-        name: 'First Section',
-        value: 'First Section Value',
-        showTitle: true,
-      },
-    ];
-
-    const initialContext = [
-      {
-        id: 'first-section',
-        name: 'First Section',
-        value: 'First Section Value',
-      },
-    ];
-
-    const context = applyDefaultConfig(initialContext);
-
-    expect(context).toEqual(expectedContext);
-  });
-});
-
 describe('createPageContentsContext', () => {
   it('should create a list of 1 of page contents objects if the solution has just 1 section', () => {
     const expectedPageContents = [
